[BB2-41] FEATURE PACK #1, #3 (#70)
* update dockerfile and appbuild to use ninja (#63)

* update dockerfile and appbuild to use ninja

* change readme to work with ninja

* Update appbuild.sh

* Update README.md

* Update README.md

* Swagger basic run (#64)

* run without auth

* run without auth

* add DefaultBearerAuthorization

---------


* Creating new receipt (#65)

* Adding structures

* Adding CalculateOperationOnReceipt function

* changeBalance will use from now userID instead of token

* deleting .vscode settings and build

* now function getBalance uses userID instead of token

* now CalculateOperationOnReceipt changes user balance

* tests for balance work on userID

* delete core.cpp & hpp (not been used)

* delete core.cpp & hpp pt2

* change name of calculating recipt

* add reciptOperations.hpp to CMake & fix error

* add receipt & implement list -> vector of items

* add conversion from answer from core to dto

* clang-formated

* update mycontroller to fix error

---------


* Recipt fix (#67)

* fix #66

* if userID not exist it return an error

* return in 14 now returns {} insted of null structure

* I changed swagger document info  and I've added new servers to which you can send requests (#68)

* I added namespaces bakcyl::api, bakcyl::database, bakcyl::core (#69)

---------

+++ appbuild.sh
@@ -1,14 +1,12 @@
 #!/bin/bash
 
-rm -rf paragon_back
+if [ -d "paragon_back" ]; then
+    rm -rf paragon_back
+fi
 
-git clone -b $BRANCH  $REPO
+git clone -b $BRANCH $REPO paragon_back
 cd paragon_back
-mkdir build
-cd build
-cmake ..
+cmake -Bbuild -GNinja
+cmake --build build --target all
+ctest --test-dir build
 cd ..
-cmake --build build
-cd build
-ctest
-cd ../..
